+++ packages/liveblocks-node/.gitignore
@@ -1,0 +1,2 @@
+# TODO: Temporary, remove after rotating keys
+src/__tests__/webhooks.test.ts
